Time functions moved to proper file outside of utils
